requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Django==2.1.1
django-qr-code==1.0.0
djangorestframework==3.8.2
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
pycparser==2.19
pytz==2018.5
qrcode==6.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
